Article Review AI "Podcast"
Review of:
How the Presumption of Atheism, by way of Semiotic Square of Opposition, leads to a Semantic Collapse

Steve McRae

Automatic AI "podcast" "Richard's Reflections"
https://www.academia.edu/80085203/How_the_Presumption_of_Atheism_by_way_of_Semiotic_Square_of_Opposition_leads_to_a_Semantic_Collapse

Sorry about the background noise. I will clean it up later, maybe. You
can probably hear the summary yourself, although as it an "AI Summary",
theoretically they could generate a new one each time or for each user.

I recorded it for posterity, although it is a little weird. I'm not sure
why the robot is allowed to pose as human, it calls itself "Richard
Price" which oddly enough, was the name of the guy who first published Bayes' work.